requirements.txt
Django==4.2.7
selenium==4.15.0
requests==2.31.0
urllib3==2.1.0
Pillow==10.0.1